chore(build): move link local config into npm script
Not sure whether its a good idea  to reference lib package(s)
both in the dependency and devDependency section of
package.json to configure.

To avoid possible unforseen problem, we move the library
to link local from the devDependency section into the npm
script.

A consequence is that we now need to rename the lib folder